Fix bug:string is not escaped.
In the original, string within '<' and/or '&' and/or ' ' is not escaped, so if it happend, xml parsing will be error, now the string will be escaped if necessary.

ps:In the original, lua table which is an array will be coverted to struct type, now a lua table which is an array will be coverted to array type.